# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AqarGo API",
        "description": "AqarGo real estate platform API documentation",
        "contact": {
            "email": "support@aqargo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/report/index": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "List all reports (admin)",
                "operationId": "7aaabb10d9e0692e53df28f3f08e06b0",
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "500": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/report/create": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Submit a report against an ad",
                "operationId": "a5367b1dc6eb8808a6204c35304d18ac",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ad_id"
                                ],
                                "properties": {
                                    "ad_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "spam",
                                        "enum": [
                                            "sexual_content",
                                            "harassment",
                                            "spam",
                                            "hate_speech",
                                            "violence",
                                            "scam",
                                            "fake_information",
                                            "other"
                                        ]
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "This ad seems fraudulent"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "422": {
                        "$ref": "#/components/schemas/ValidationErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/report/show/{id}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get a report by ID",
                "operationId": "7627dca2e567d7f031c660b51aa175f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "404": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/report/showAdReports": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get all reports for a specific ad",
                "operationId": "c7d4123740e681999e658ce3094892e2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ad_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "500": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/report/delete/{id}": {
            "delete": {
                "tags": [
                    "Reports"
                ],
                "summary": "Delete a report (admin)",
                "operationId": "7508ddda5116084541a92c4a171941ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "500": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 1
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 0
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 0
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation Error."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your JWT token"
            }
        }
    },
...
}